requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tokenizers==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.17.0
